Update `getContract` function
To be able to create a contract instance with a signer to sign
transaction. Also add `useEffect` to the `ThresholdContext` to update
the provider if the user connects to a wallet.
