+++ src/threshold-ts/index.ts
@@ -4,11 +4,15 @@ import { IStaking, Staking } from "./staking"
 import { ThresholdConfig } from "./types"
 
 export class Threshold {
-  multicall: IMulticall
-  staking: IStaking
-  multiAppStaking: MultiAppStaking
+  multicall!: IMulticall
+  staking!: IStaking
+  multiAppStaking!: MultiAppStaking
 
   constructor(config: ThresholdConfig) {
+    this._initialize(config)
+  }
+
+  private _initialize = (config: ThresholdConfig) => {
     this.multicall = new Multicall(config.ethereum)
     this.staking = new Staking(config.ethereum)
     this.multiAppStaking = new MultiAppStaking(
@@ -16,5 +20,9 @@ export class Threshold {
       this.multicall,
       config.ethereum
     )
+  }
+
+  updateConfig = (config: ThresholdConfig) => {
+    this._initialize(config)
   }
 }
