+++ src/threshold-ts/utils/contract.ts
@@ -1,9 +1,35 @@
+import { JsonRpcSigner, Web3Provider } from "@ethersproject/providers"
 import { Contract, ContractInterface, providers, Signer } from "ethers"
+import { getAddress, isAddressZero } from "./address"
+
+// account is not optional
+export function getSigner(
+  library: Web3Provider,
+  account: string
+): JsonRpcSigner {
+  return library.getSigner(account).connectUnchecked()
+}
+
+// account is optional
+function getProviderOrSigner(
+  library: Web3Provider,
+  account?: string
+): Web3Provider | JsonRpcSigner {
+  return account ? getSigner(library, account) : library
+}
 
 export const getContract = (
   address: string,
   abi: ContractInterface,
-  providerOrSigner: providers.Provider | Signer | undefined
+  providerOrSigner: providers.Provider | Signer | undefined,
+  account?: string
 ) => {
-  return new Contract(address, abi, providerOrSigner)
+  if (!getAddress(address) || isAddressZero(address)) {
+    throw Error(`Invalid 'address' parameter '${address}'.`)
+  }
+  return new Contract(
+    address,
+    abi,
+    getProviderOrSigner(providerOrSigner as any, account) as any
+  )
 }
