+++ src/utils/getContract.ts
@@ -1,41 +1,5 @@
-import { getAddress } from "@ethersproject/address"
-import { AddressZero } from "@ethersproject/constants"
-import { Contract } from "@ethersproject/contracts"
-import { JsonRpcSigner, Web3Provider } from "@ethersproject/providers"
-import { isAddressZero } from "../web3/utils"
-
-// account is not optional
-export function getSigner(
-  library: Web3Provider,
-  account: string
-): JsonRpcSigner {
-  return library.getSigner(account).connectUnchecked()
-}
-
-// account is optional
-function getProviderOrSigner(
-  library: Web3Provider,
-  account?: string
-): Web3Provider | JsonRpcSigner {
-  return account ? getSigner(library, account) : library
-}
-
-export function getContract(
-  address: string,
-  ABI: any,
-  library: Web3Provider,
-  account?: string
-): Contract {
-  if (!getAddress(address) || isAddressZero(address)) {
-    throw Error(`Invalid 'address' parameter '${address}'.`)
-  }
-
-  return new Contract(
-    address,
-    ABI,
-    getProviderOrSigner(library, account) as any
-  )
-}
+import { AddressZero } from "../web3/utils/address"
+export { getSigner, getContract } from "../threshold-ts/utils/contract"
 
 export function getContractAddressFromTruffleArtifact(
   truffleArtifact: { networks: { [chainID: string]: { address: string } } },
